+++ library/src/androidTest/java/com/evernote/android/job/Platform21Test.java
@@ -1,11 +1,14 @@
 package com.evernote.android.job;
 
+import android.os.Build;
 import android.support.test.filters.LargeTest;
 import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import static org.junit.Assume.assumeTrue;
 
 /**
  * @author rwondratschek
@@ -19,6 +22,8 @@ public class Platform21Test {
 
     @Test(expected = IllegalStateException.class)
     public void test100DistinctJobsLimit() {
+        assumeTrue(Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1);
+
         for (int i = 0; i < 130; i++) {
             new JobRequest.Builder("tag")
                     .setExecutionWindow(30_000, 40_000)